FAR patch 01260.Mix Дата              : 26.02.2002 Сделал            : Valentin Skirdin Описание          : Mix Измененные файлы  : config.cpp farrtl.cpp history.cpp plist.cpp Новые файлы       : Удаленные файлы   : Состав            : 01260.Mix.txt                     config.cpp.1260.diff                     farrtl.cpp.1260.diff                     history.cpp.1260.diff                     plist.cpp.1260.diff Основан на патче  : 1259 Дополнение        :
1. Про хистори.
   AAB>  При удалении хистори (Alt+F11, Alt+F12, в диалогах итд),
   AAB>  при пустом списке, вопрос на подтверждение удаления задавать не стоит.

2. RestoreCP ставим по умолчанию в 0.

3. Баги в VC-реализации функций ftell64 и fseek64

4. Активизация процесса с подачи MY.

+++ far/history.cpp
@@ -5,10 +5,12 @@ history.cpp
 
 */
 
-/* Revision: 1.20 25.01.2002 $ */
+/* Revision: 1.21 26.02.2002 $ */
 
 /*
 Modify:
+  26.02.2002 SVS
+    ! Äëÿ ïóñòîãî ñïèñêà (Items > 1) íè÷åãî íå äåëàåì
   25.01.2002 SVS
     - ìàààààààààààñññññññññññññññññääääääääääääààààààààààààààààààéééééééééé!!!!
   25.01.2002 SVS
@@ -450,14 +452,15 @@ int History::Select(char *Title,char *HelpTopic,char *Str,int &Type,char *ItemTi
         /* $ 23.07.2001 VVM
           + Ñïðîñèòü ïîäòâåðæäåíèå ïåðåä óäàëåíèåì */
         {
-          if (!Opt.Confirm.HistoryClear ||
+          if(HistoryMenu.GetItemCount() > 1 &&
+             (!Opt.Confirm.HistoryClear ||
               (Opt.Confirm.HistoryClear &&
               Message(MSG_WARNING,2,
                    MSG((History::TypeHistory==HISTORYTYPE_CMD?MHistoryTitle:
                         (History::TypeHistory==HISTORYTYPE_FOLDER?MFolderHistoryTitle:
                         MViewHistoryTitle))),
                    MSG(MHistoryClear),
-                   MSG(MClear),MSG(MCancel))==0))
+                   MSG(MClear),MSG(MCancel))==0)))
           {
             FreeHistory();
             DeleteRegValue(RegKey,"Lines");
